# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BitaSoft Api Documentation",
        "description": "BitaSoft Api Documentation",
        "contact": {
            "name": "Bita Soft",
            "email": "bitasoft22@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/v1/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of Users",
                "description": "Get list of Users",
                "operationId": "ac5e3db32574f697d2620c47e7f40965",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "the page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "order accepts 'asc' or 'desc'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Store Users in DB",
                "description": "Store Users in DB",
                "operationId": "32a373267b9e6073ecec118740dd58ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstName",
                                    "lastName"
                                ],
                                "properties": {
                                    "firstName": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/v1/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Users Detail",
                "description": "Get Users Detail",
                "operationId": "4056162dfcb710509d75d2cd7fddf1c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Users",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update Users in DB",
                "description": "Update Users in DB",
                "operationId": "d6b63fea683122b344ada4ea3bc73284",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Users",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstName",
                                    "lastName"
                                ],
                                "properties": {
                                    "firstName": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete Users",
                "description": "Delete Users",
                "operationId": "661d8cfd72df9a89a152cdf32222fb3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Users",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}